+++ src/main/java/mdlaf/components/spinner/MaterialSpinnerUI.java
@@ -1,18 +1,18 @@
 /**
  * MIT License
- *
+ * <p>
  * Copyright (c) 2018-2020 atharva washimkar, Vincenzo Palazzo vincenzopalazzo1996@gmail.com
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -52,17 +52,21 @@ public class MaterialSpinnerUI extends BasicSpinnerUI {
         super.installUI(c);
 
         JSpinner spinner = (JSpinner) c;
-        spinner.setOpaque(false);
+        //spinner.setOpaque(false);
         spinner.setFont(UIManager.getFont("Spinner.font"));
         spinner.setBackground(UIManager.getColor("Spinner.background"));
         spinner.setForeground(UIManager.getColor("Spinner.foreground"));
-        spinner.setBorder(UIManager.getBorder("Spinner.border"));
+        spinner.getEditor().setBorder(UIManager.getBorder("Spinner.border"));
         //c.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
     }
 
     @Override
     public void uninstallUI(JComponent c) {
         //c.setCursor(null);
+        spinner.setFont(null);
+        spinner.setBackground(null);
+        spinner.setForeground(null);
+        spinner.getEditor().setBorder(null);
 
         super.uninstallUI(c);
     }
@@ -89,9 +93,9 @@ public class MaterialSpinnerUI extends BasicSpinnerUI {
         } else {
             button = new BasicArrowButton(SwingConstants.NORTH);
         }
-       this.configureLocalButton(button);
+        this.configureLocalButton(button);
         installNextButtonListeners(button);
-        button.setBorder(BorderFactory.createLineBorder(button.getBackground()));
+        //button.setBorder(BorderFactory.createLineBorder(button.getBackground()));
         this.upArrowButton = button;
         return button;
     }
@@ -107,17 +111,20 @@ public class MaterialSpinnerUI extends BasicSpinnerUI {
         }
         this.configureLocalButton(button);
         installPreviousButtonListeners(button);
-        button.setBorder (BorderFactory.createLineBorder(button.getBackground()));
+        //button.setBorder(BorderFactory.createLineBorder(button.getBackground()));
         this.downArrowButton = button;
         return button;
     }
 
-    protected void configureLocalButton(JButton arrowButton){
+    protected void configureLocalButton(JButton arrowButton) {
+        if (arrowButton == null) {
+            return;
+        }
         arrowButton.setOpaque(true);
         arrowButton.setBackground(UIManager.getColor("Spinner.arrowButtonBackground"));
-        if(UIManager.getBoolean("Spinner.mouseHoverEnabled")){
-            arrowButton.addMouseListener(MaterialUIMovement.getMovement(arrowButton, UIManager.getColor ("Spinner.mouseHoverColor")));
+        if (UIManager.getBoolean("Spinner.mouseHoverEnabled")) {
+            arrowButton.addMouseListener(MaterialUIMovement.getMovement(arrowButton, UIManager.getColor("Spinner.mouseHoverColor")));
         }
-        arrowButton.setBorder (BorderFactory.createLineBorder(arrowButton.getBackground()));
+        arrowButton.setBorder(BorderFactory.createLineBorder(arrowButton.getBackground()));
     }
 }