+++ src/main/java/mdlaf/components/table/MaterialTableHeaderCellRenderer.java
@@ -1,18 +1,18 @@
 /**
  * MIT License
- *
+ * <p>
  * Copyright (c) 2018-2020 atharva washimkar, Vincenzo Palazzo vincenzopalazzo1996@gmail.com
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -36,17 +36,17 @@ import java.awt.Component;
  */
 public class MaterialTableHeaderCellRenderer extends DefaultTableCellRenderer {
 
-	@Override
-	public Component getTableCellRendererComponent (JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
-		JComponent component = (JComponent) super.getTableCellRendererComponent (table, value, isSelected, hasFocus, row, column);
-		component.setBorder (UIManager.getBorder ("TableHeader.cellBorder"));
-		component.setFont (UIManager.getFont ("TableHeader.font"));
-		component.setBackground (UIManager.getColor ("TableHeader.background"));
-		component.setForeground(UIManager.getColor("TableHeader.foreground"));
+    @Override
+    public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
+        super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
+        setBorder(UIManager.getBorder("TableHeader.cellBorder"));
+        setFont(UIManager.getFont("TableHeader.font"));
+        setBackground(UIManager.getColor("TableHeader.background"));
+        setForeground(UIManager.getColor("TableHeader.foreground"));
 
-		this.setHorizontalAlignment (SwingConstants.CENTER);
-		this.setVerticalAlignment (SwingConstants.CENTER);
+        this.setHorizontalAlignment(SwingConstants.CENTER);
+        this.setVerticalAlignment(SwingConstants.CENTER);
 
-		return component;
-	}
+        return this;
+    }
 }