+++ src/test/java/integration/gui/mock/TableModelSecondPanel.java
@@ -23,34 +23,32 @@
  */
 package integration.gui.mock;
 
+
 import javax.swing.table.AbstractTableModel;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * @author https://github.com/vincenzopalazzo
  */
 public class TableModelSecondPanel extends AbstractTableModel {
 
-    private String[] columnNames = {"First Name",
-            "Last Name",
-            "Sport",
-            "# of Years",
-            "Vegetarian"};
-    private Object[][] data = {
-            {"Kathy", "Smith",
-                    "Snowboarding", new Integer(5), new Boolean(false)},
-            {"John", "Doe",
-                    "Rowing", new Integer(3), new Boolean(true)},
-            {"Sue", "Black",
-                    "Knitting", new Integer(2), new Boolean(false)},
-            {"Jane", "White",
-                    "Speed reading", new Integer(20), new Boolean(true)},
-            {"Joe", "Brown",
-                    "Pool", new Integer(10), new Boolean(false)}
-    };
+    private boolean DEBUG = true;
 
-    public final Object[] longValues = {"Jane", "Kathy",
-            "None of the above",
-            new Integer(20), Boolean.TRUE};
+    private String[] columnNames = { "First Name", "Last Name", "Sport",
+            "# of Years", "Vegetarian" };
+    private List<String> columNameStringList = new ArrayList<>(Arrays.asList(columnNames));
+    private Object[][] data = {
+            { "Mary", "Campione", "Snowboarding", new Integer(5),
+                    new Boolean(false) },
+            { "Alison", "Huml", "Rowing", new Integer(3), new Boolean(true) },
+            { "Kathy", "Walrath", "Knitting", new Integer(2),
+                    new Boolean(false) },
+            { "Sharon", "Zakhour", "Speed reading", new Integer(20),
+                    new Boolean(true) },
+            { "Philip", "Milne", "Pool", new Integer(10),
+                    new Boolean(false) } };
 
     public int getColumnCount() {
         return columnNames.length;
@@ -60,48 +58,46 @@ public class TableModelSecondPanel extends AbstractTableModel {
         return data.length;
     }
 
-    @Override
-    public String getColumnName(int column) {
-        return columnNames[column];
+    public String getColumnName(int col) {
+        return columNameStringList.get(col);
     }
 
     public Object getValueAt(int row, int col) {
         return data[row][col];
     }
 
-    /*
-     * JTable uses this method to determine the default renderer/
-     * editor for each cell.  If we didn't implement this method,
-     * then the last column would contain text ("true"/"false"),
-     * rather than a check box.
-     */
     public Class getColumnClass(int c) {
         return getValueAt(0, c).getClass();
     }
 
-    /*
-     * Don't need to implement this method unless your table's
-     * editable.
-     */
-    public boolean isCellEditable(int row, int col) {
-        //Note that the data/cell address is constant,
-        //no matter where the cell appears onscreen.
-        if (col < 2) {
-            return false;
-        } else {
-            return true;
+    public void setValueAt(Object value, int row, int col) {
+        if (DEBUG) {
+            System.out.println("Setting value at " + row + "," + col
+                    + " to " + value + " (an instance of "
+                    + value.getClass() + ")");
+        }
+
+        data[row][col] = value;
+        fireTableCellUpdated(row, col);
+
+        if (DEBUG) {
+            System.out.println("New value of data:");
+            printDebugData();
         }
     }
 
-    /*
-     * Don't need to implement this method unless your table's
-     * data can change.
-     */
-    public void setValueAt(Object value, int row, int col) {
-        data[row][col] = value;
-        fireTableCellUpdated(row, col);
+    private void printDebugData() {
+        int numRows = getRowCount();
+        int numCols = getColumnCount();
+
+        for (int i = 0; i < numRows; i++) {
+            System.out.print("    row " + i + ":");
+            for (int j = 0; j < numCols; j++) {
+                System.out.print("  " + data[i][j]);
+            }
+            System.out.println();
+        }
+        System.out.println("--------------------------");
     }
-
-
 }